+++ students/2016-09-04/Paolo/managers02/gestionale_su_db.py
@@ -4,6 +4,7 @@ import sys
 from collections import defaultdict
 import sqlite3
 import os
+import sqlitedatabase as sqlmanager
 
 PEOPLE = []
 
@@ -32,36 +33,6 @@ def main():
     print_on_screen(PEOPLE)
     
     
-class SqliteDatabase(object):
-    def __init__(self, file_name):
-        self.file_name = file_name
-        
-        self.conn = sqlite3.connect(file_name)
-        self.conn.row_factory = sqlite3.Row
-
-        cursor = self.conn.cursor()
-        
-        if not os.path.exists(file_name):
-            cursor.execute("CREATE TABLE people (name VARCHAR(64), city VARCHAR(32), salary INTEGER);")
-            
-    def write(self, list):
-        cursor = self.conn.cursor()
-        
-        for row in list:
-            t = (row["name"], row["city"], row["salary"])
-            cursor.execute('INSERT INTO people VALUES (?,?,?)', t)
-            
-    def read_all(self):
-        cursor = self.conn.cursor()
-        cursor.execute("SELECT * FROM people")
-        return cursor.fetchall()
-        
-    def commit(self):
-        self.conn.commit()
-        
-    def close(self):
-        self.conn.close()
-    
 def save_on_db(list_input, database_manager):
     database_manager.write(list_input)
     database_manager.commit()
@@ -76,7 +47,7 @@ def read_from_db(database_manager):
 def main_and_save(input_args):
     main()
     
-    database_manager = SqliteDatabase("mio_db.db")
+    database_manager = sqlmanager.SqliteDatabase("mio_db.db")
     save_on_db(PEOPLE, database_manager)
     read_from_db(database_manager)
